+++ routers/home.js
@@ -1,5 +1,4 @@
 const { Device } = require('../controllers/engines/router');
-const { _M_ } = require('../controllers/middleware/plugins');
 const home = require('../controllers/routers/home');
 
 
@@ -10,25 +9,24 @@ const home = require('../controllers/routers/home');
 const DashboardRouter = new Device([{
     route:          '/',
     controller:     home.main,
-    settings:       { title: 'Dashboard', template: './__root__/' },
+    setting:        { title: 'Dashboard', template: './__root__/' },
 }, {
     route:          '/profile',
     controller:     home.profile,
-    settings:       { title: 'Profile', template: './__root__/home/profile/info' },
+    setting:        { title: 'Profile', template: './__root__/home/profile/info' },
 }, {
     route:          '/profile/edit',
     controller:     home.profile_editor,
-    settings:       { title: 'Edit Your Profile', template: './__root__/home/profile/profile_editor' },
+    setting:        { title: 'Edit Your Profile', template: './__root__/home/profile/profile_editor' },
 }, {
     route:          '/security',
     controller:     home.security,
-    settings:       { title: 'Change Password', template: './__root__/home/profile/security' },
+    setting:        { title: 'Change Password', template: './__root__/home/profile/security' },
 }]);
 
 
 // pre-used middleware
-DashboardRouter.pre([_M_.isSignedIn, _M_.setTitleTag('Account')]);
-
+DashboardRouter.setting = { title: 'Account', authenticated : true };
 
 
 // router exports